--- attic/playbooks/play1.yml
@@ -1,12 +1,0 @@
----
-- hosts: localhost
-  vars_files:
-    - vars.yml
-  tasks:
-  - name: Get HMC Webservice API info.
-    zhmc_api:
-      auth_hmc: "{{ hmc }}"
-    register: result
-
-  - debug: var=result
-
