+++ collections/requirements.yml
@@ -1,0 +1,2 @@
+collections:
+  - ibm.zhmc
