+++ dev-requirements.txt
@@ -64,7 +64,7 @@ dparse>=0.4.1; python_version >= '3.4'
 tox>=2.5.0
 
 # Sphinx (no imports, invoked via sphinx-build script):
-# Keep in sync with rtd-requirements.txt
+# Keep in sync with rtd-requirements.yml
 Sphinx>=1.7.6
 sphinx-git>=10.1.1
 GitPython>=2.1.1
