+++ docs/conf.py
@@ -13,6 +13,9 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
+from __future__ import (absolute_import, division, print_function)
+__metaclass__ = type
+
 import sys
 import os
 import re
@@ -33,15 +36,15 @@ needs_sphinx = '1.5'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-#    'sphinx.ext.autodoc',
-#    'sphinx.ext.napoleon',
+    # 'sphinx.ext.autodoc',
+    # 'sphinx.ext.napoleon',
     'sphinx.ext.intersphinx',
     'sphinx.ext.extlinks',
-#    'sphinx.ext.todo',
-#    'sphinx.ext.coverage',
-#    'sphinx.ext.viewcode',   # disabed, raises anexception
-#    'sphinx.ext.ifconfig',
-#    'sphinx_git',            # requires 'sphinx-git' Python package
+    # 'sphinx.ext.todo',
+    # 'sphinx.ext.coverage',
+    # 'sphinx.ext.viewcode',   # disabled, raises an exception
+    # 'sphinx.ext.ifconfig',
+    # 'sphinx_git',            # requires 'sphinx-git' Python package
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -74,6 +77,7 @@ _short_description = u'Ansible modules for IBM Z and LinuxONE'
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 
+
 def package_version(package_name):
     """Return package version string by reading `filename` and retrieving its
        module-global variable `varnam`."""
@@ -87,6 +91,7 @@ def package_version(package_name):
             version = m.group(1)
             break
     return version
+
 
 # The short X.Y version.
 # Note: We use the full version in both cases (e.g. 'M.N.U' or 'M.N.U.dev0').
@@ -104,9 +109,9 @@ language = None
 
 # There are two options for replacing |today|: either, you set today to some
 # non-false value, then it is used:
-#today = ''
+# today = ''
 # Else, today_fmt is used as the format for a strftime call.
-#today_fmt = '%B %d, %Y'
+# today_fmt = '%B %d, %Y'
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -122,20 +127,20 @@ add_function_parentheses = True
 
 # If true, the current module name will be prepended to all description
 # unit titles (such as .. function::).
-#add_module_names = True
+# add_module_names = True
 
 # If true, sectionauthor and moduleauthor directives will be shown in the
 # output. They are ignored by default.
-#show_authors = False
+# show_authors = False
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = 'sphinx'
 
 # A list of ignored prefixes for module index sorting.
-#modindex_common_prefix = []
+# modindex_common_prefix = []
 
 # If true, keep warnings as "system message" paragraphs in the built documents.
-#keep_warnings = False
+# keep_warnings = False
 
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = True
@@ -149,11 +154,11 @@ napoleon_google_docstring = True
 # Enable support for NumPy style docstrings. Defaults to True.
 napoleon_numpy_docstring = False
 
-# Include private members (like _membername). False to fall back to Sphinx’s
+# Include private members (like _membername). False to fall back to Sphinx's
 # default behavior. Defaults to False.
 napoleon_include_private_with_doc = False
 
-# Include special members (like __membername__). False to fall back to Sphinx’s
+# Include special members (like __membername__). False to fall back to Sphinx's
 # default behavior. Defaults to True.
 napoleon_include_special_with_doc = True
 
@@ -205,23 +210,23 @@ html_theme_options = {
 }
 
 # Add any paths that contain custom themes here, relative to this directory.
-#html_theme_path = []
+# html_theme_path = []
 
 # The name for this set of Sphinx documents.  If not defined, it defaults to
 # "<project> v<release> documentation".
-#html_title = None
+# html_title = None
 
 # A shorter title for the navigation bar.  Default is the same as html_title.
-#html_short_title = 'ld'
+# html_short_title = 'ld'
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
-#html_logo = None
+# html_logo = None
 
 # The name of an image file (relative to this directory) to use as a favicon of
-# the docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
-# pixels large.
-#html_favicon = None
+# the docs.  This file should be a Windows icon file (.ico) being 16x16
+# or 32x32 pixels large.
+# html_favicon = None
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -235,104 +240,103 @@ html_extra_path = ['_extra']
 
 # If not '', a 'Last updated on:' timestamp is inserted at every page bottom,
 # using the given strftime format.
-#html_last_updated_fmt = '%b %d, %Y'
+# html_last_updated_fmt = '%b %d, %Y'
 
 # If true, SmartyPants will be used to convert quotes and dashes to
 # typographically correct entities.
-#html_use_smartypants = True
+# html_use_smartypants = True
 
 # Custom sidebar templates, maps document names to template names.
-#html_sidebars = {}
+# html_sidebars = {}
 
 # Additional templates that should be rendered to pages, maps page names to
 # template names.
-#html_additional_pages = {}
+# html_additional_pages = {}
 
 # If false, no module index is generated.
-#html_domain_indices = True
+# html_domain_indices = True
 
 # If false, no index is generated.
-#html_use_index = True
+# html_use_index = True
 
 # If true, the index is split into individual pages for each letter.
-#html_split_index = False
+# html_split_index = False
 
 # If true, links to the reST sources are added to the pages.
-#html_show_sourcelink = True
+# html_show_sourcelink = True
 
 # If true, "Created using Sphinx" is shown in the HTML footer. Default is True.
-#html_show_sphinx = True
+# html_show_sphinx = True
 
 # If true, "(C) Copyright ..." is shown in the HTML footer. Default is True.
-#html_show_copyright = True
+# html_show_copyright = True
 
 # If true, an OpenSearch description file will be output, and all pages will
 # contain a <link> tag referring to it.  The value of this option must be the
 # base URL from which the finished HTML is served.
-#html_use_opensearch = ''
+# html_use_opensearch = ''
 
 # This is the file name suffix for HTML files (e.g. ".xhtml").
-#html_file_suffix = None
+# html_file_suffix = None
 
 # Language to be used for generating the HTML full-text search index.
 # Sphinx supports the following languages:
 #   'da', 'de', 'en', 'es', 'fi', 'fr', 'hu', 'it', 'ja'
 #   'nl', 'no', 'pt', 'ro', 'ru', 'sv', 'tr'
-#html_search_language = 'en'
+# html_search_language = 'en'
 
 # A dictionary with options for the search language support, empty by default.
 # Now only 'ja' uses this config value
-#html_search_options = {'type': 'default'}
+# html_search_options = {'type': 'default'}
 
 # The name of a javascript file (relative to the configuration directory) that
 # implements a search results scorer. If empty, the default will be used.
-#html_search_scorer = 'scorer.js'
+# html_search_scorer = 'scorer.js'
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = project+'_doc'
+htmlhelp_basename = project + '_doc'
 
 # -- Options for LaTeX output ---------------------------------------------
 
 latex_elements = {
-# The paper size ('letterpaper' or 'a4paper').
-'papersize': 'a4paper',
+    # The paper size ('letterpaper' or 'a4paper').
+    'papersize': 'a4paper',
+    # The font size ('10pt', '11pt' or '12pt').
+    # 'pointsize': '10pt',
 
-# The font size ('10pt', '11pt' or '12pt').
-#'pointsize': '10pt',
+    # Additional stuff for the LaTeX preamble.
+    # 'preamble': '',
 
-# Additional stuff for the LaTeX preamble.
-#'preamble': '',
-
-# Latex figure (float) alignment
-#'figure_align': 'htbp',
+    # Latex figure (float) alignment
+    # 'figure_align': 'htbp',
 }
 
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, project+'.tex', _short_description, author, 'manual'),
+    (master_doc, project + '.tex', _short_description, author, 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
 # the title page.
-#latex_logo = None
+# latex_logo = None
 
 # For "manual" documents, if this is true, then toplevel headings are parts,
 # not chapters.
-#latex_use_parts = False
+# latex_use_parts = False
 
 # If true, show page references after internal links.
-#latex_show_pagerefs = False
+# latex_show_pagerefs = False
 
 # If true, show URL addresses after external links.
-#latex_show_urls = False
+# latex_show_urls = False
 
 # Documents to append as an appendix to all manuals.
-#latex_appendices = []
+# latex_appendices = []
 
 # If false, no module index is generated.
-#latex_domain_indices = True
+# latex_domain_indices = True
 
 
 # -- Options for manual page output ---------------------------------------
@@ -344,7 +348,7 @@ man_pages = [
 ]
 
 # If true, show URL addresses after external links.
-#man_show_urls = False
+# man_show_urls = False
 
 
 # -- Options for Texinfo output -------------------------------------------
@@ -359,16 +363,16 @@ texinfo_documents = [
 ]
 
 # Documents to append as an appendix to all manuals.
-#texinfo_appendices = []
+# texinfo_appendices = []
 
 # If false, no module index is generated.
-#texinfo_domain_indices = True
+# texinfo_domain_indices = True
 
 # How to display URL addresses: 'footnote', 'no', or 'inline'.
-#texinfo_show_urls = 'footnote'
+# texinfo_show_urls = 'footnote'
 
 # If true, do not generate a @detailmenu in the "Top" node's menu.
-#texinfo_no_detailmenu = False
+# texinfo_no_detailmenu = False
 
 
 # -- Options for autodoc extension ----------------------------------------
@@ -377,10 +381,10 @@ texinfo_documents = [
 
 # Selects what content will be inserted into a class description.
 # The possible values are:
-#   "class" - Only the class’ docstring is inserted. This is the default.
-#   "both"  - Both the class’ and the __init__ method’s docstring are
+#   "class" - Only the class' docstring is inserted. This is the default.
+#   "both"  - Both the class' and the __init__ method's docstring are
 #             concatenated and inserted.
-#   "init"  - Only the __init__ method’s docstring is inserted.
+#   "init"  - Only the __init__ method's docstring is inserted.
 # In all cases, the __init__ method is still independently rendered as a
 # special method, e.g. when the :special-members: option is set.
 autoclass_content = "both"
@@ -401,7 +405,7 @@ autodoc_default_flags = []
 # Functions imported from C modules cannot be introspected, and therefore the
 # signature for such functions cannot be automatically determined. However, it
 # is an often-used convention to put the signature into the first line of the
-# function’s docstring.
+# function's docstring.
 # If this boolean value is set to True (which is the default), autodoc will
 # look at the first line of the docstring for functions and methods, and if it
 # looks like a signature, use the line as the signature and remove it from the
@@ -433,9 +437,9 @@ autodoc_mock_imports = []
 #       to datatypes of function parameters can be controlled.
 #
 intersphinx_mapping = {
-  'py': ('https://docs.python.org/2/', None), # agnostic to Python version
-  'py2': ('https://docs.python.org/2', None), # specific to Python 2
-  'py3': ('https://docs.python.org/3', None), # specific to Python 3
+    'py': ('https://docs.python.org/2/', None),  # agnostic to Python version
+    'py2': ('https://docs.python.org/2', None),  # specific to Python 2
+    'py3': ('https://docs.python.org/3', None),  # specific to Python 3
 }
 
 intersphinx_cache_limit = 5
@@ -480,7 +484,9 @@ intersphinx_cache_limit = 5
 # results in the link caption "this issue".
 
 extlinks = {
-  'nbview': ('http://nbviewer.jupyter.org/github/zhmcclient/python-zhmcclient/blob/master/docs/notebooks/%s', ''),
-  'nbdown': ('https://github.com/zhmcclient/python-zhmcclient/raw/master/docs/notebooks/%s', '')
+    'nbview': ('http://nbviewer.jupyter.org/github/'
+               'zhmcclient/python-zhmcclient'
+               '/blob/master/docs/notebooks/%s', ''),
+    'nbdown': ('https://github.com/zhmcclient/python-zhmcclient/raw/master'
+               '/docs/notebooks/%s', '')
 }
-
