+++ plugins/modules/zhmc_cpc.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/python
 # Copyright 2018 IBM Corp. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,15 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import, print_function
-
-import logging
-from ansible.module_utils.basic import AnsibleModule
-import requests.packages.urllib3
-import zhmcclient
-
-from zhmc_ansible_modules.utils import log_init, Error, ParameterError, \
-    get_hmc_auth, get_session, to_unicode, process_normal_property
+from __future__ import (absolute_import, division, print_function)
+__metaclass__ = type
 
 # For information on the format of the ANSIBLE_METADATA, DOCUMENTATION,
 # EXAMPLES, and RETURN strings, see
@@ -45,8 +38,8 @@ description:
   - Updates the properties of a CPC.
 notes:
 author:
-  - Andreas Maier (@andy-maier, maiera@de.ibm.com)
-  - Andreas Scheuring (@scheuran, scheuran@de.ibm.com)
+  - Andreas Maier (@andy-maier)
+  - Andreas Scheuring (@scheuran)
 requirements:
   - Network access to HMC
   - zhmcclient >=0.20.0
@@ -101,7 +94,7 @@ options:
          defined as writeable in the data model for CPC resources."
     type: dict
     required: false
-    default: No property changes.
+    default: null
   log_file:
     description:
       - "File path of a log file to which the logic flow of this module as well
@@ -116,7 +109,7 @@ options:
          If provided, it will be used instead of connecting to a real HMC. This
          is used for testing purposes only."
     required: false
-    default: Real HMC will be used.
+    type: raw
 """
 
 EXAMPLES = """
@@ -181,6 +174,28 @@ cpc:
       ],
     })
 """
+
+import logging  # noqa: E402
+import traceback  # noqa: E402
+from ansible.module_utils.basic import AnsibleModule  # noqa: E402
+
+from ..module_utils.common import log_init, Error, ParameterError, \
+    get_hmc_auth, get_session, to_unicode, process_normal_property, \
+    missing_required_lib  # noqa: E402
+
+try:
+    import requests.packages.urllib3
+    IMP_URLLIB3 = True
+except ImportError:
+    IMP_URLLIB3 = False
+    IMP_URLLIB3_ERR = traceback.format_exc()
+
+try:
+    import zhmcclient
+    IMP_ZHMCCLIENT = True
+except ImportError:
+    IMP_ZHMCCLIENT = False
+    IMP_ZHMCCLIENT_ERR = traceback.format_exc()
 
 # Python logger name for this module
 LOGGER_NAME = 'zhmc_cpc'
@@ -262,15 +277,17 @@ def process_properties(cpc, params):
 
         if not allowed:
             raise ParameterError(
-                "CPC property {!r} specified in the 'properties' module "
+                "CPC property {0!r} specified in the 'properties' module "
                 "parameter cannot be updated.".format(prop_name))
 
         # Process a normal (= non-artificial) property
         _create_props, _update_props, _stop = process_normal_property(
             prop_name, ZHMC_CPC_PROPERTIES, input_props, cpc)
         update_props.update(_update_props)
-        assert not _create_props
-        assert _stop is False
+        if _create_props:
+            raise AssertionError()
+        if _stop:
+            raise AssertionError()
 
     return update_props
 
@@ -392,19 +409,29 @@ def main():
         state=dict(required=True, type='str', choices=['set', 'facts']),
         properties=dict(required=False, type='dict', default={}),
         log_file=dict(required=False, type='str', default=None),
-        faked_session=dict(required=False, type='object'),
+        faked_session=dict(required=False, type='raw'),
     )
 
     module = AnsibleModule(
         argument_spec=argument_spec,
         supports_check_mode=True)
 
+    if not IMP_URLLIB3:
+        module.fail_json(msg=missing_required_lib("requests"),
+                         exception=IMP_URLLIB3_ERR)
+
+    requests.packages.urllib3.disable_warnings()
+
+    if not IMP_ZHMCCLIENT:
+        module.fail_json(msg=missing_required_lib("zhmcclient"),
+                         exception=IMP_ZHMCCLIENT_ERR)
+
     log_file = module.params['log_file']
     log_init(LOGGER_NAME, log_file)
 
     _params = dict(module.params)
     del _params['hmc_auth']
-    LOGGER.debug("Module entry: params: {!r}".format(_params))
+    LOGGER.debug("Module entry: params: %r", _params)
 
     try:
 
@@ -414,21 +441,17 @@ def main():
         # These exceptions are considered errors in the environment or in user
         # input. They have a proper message that stands on its own, so we
         # simply pass that message on and will not need a traceback.
-        msg = "{}: {}".format(exc.__class__.__name__, exc)
-        LOGGER.debug(
-            "Module exit (failure): msg: {!r}".
-            format(msg))
+        msg = "{0}: {1}".format(exc.__class__.__name__, exc)
+        LOGGER.debug("Module exit (failure): msg: %r", msg)
         module.fail_json(msg=msg)
     # Other exceptions are considered module errors and are handled by Ansible
     # by showing the traceback.
 
-    LOGGER.debug(
-        "Module exit (success): changed: {!r}, cpc: {!r}".
-        format(changed, result))
+    LOGGER.debug("Module exit (success): changed: %s, cpc: %r",
+                 changed, result)
     module.exit_json(
         changed=changed, cpc=result)
 
 
 if __name__ == '__main__':
-    requests.packages.urllib3.disable_warnings()
     main()
