+++ plugins/modules/zhmc_nic.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/python
 # Copyright 2017 IBM Corp. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,16 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import, print_function
-
-import logging
-from ansible.module_utils.basic import AnsibleModule
-import requests.packages.urllib3
-import zhmcclient
-
-from zhmc_ansible_modules.utils import log_init, Error, ParameterError, \
-    wait_for_transition_completion, eq_hex, get_hmc_auth, get_session, \
-    to_unicode, process_normal_property
+from __future__ import (absolute_import, division, print_function)
+__metaclass__ = type
 
 # For information on the format of the ANSIBLE_METADATA, DOCUMENTATION,
 # EXAMPLES, and RETURN strings, see
@@ -38,20 +30,18 @@ ANSIBLE_METADATA = {
 
 DOCUMENTATION = """
 ---
-module: zhmc_virtual_function
-version_added: "0.0"
-short_description: Manages virtual functions in existing partitions
+module: zhmc_nic
+short_description: Manages NICs in existing partitions
 description:
-  - Creates, updates, and deletes virtual functions in existing partitions of a
-    CPC.
+  - Creates, updates, and deletes NICs in existing partitions of a CPC.
   - The targeted CPC must be in the Dynamic Partition Manager (DPM) operational
     mode.
 notes:
   - See also Ansible module zhmc_partition.
 author:
-  - Andreas Maier (@andy-maier, maiera@de.ibm.com)
-  - Andreas Scheuring (@scheuran, scheuran@de.ibm.com)
-  - Juergen Leopold (@leopoldjuergen, leopoldj@de.ibm.com)
+  - Andreas Maier (@andy-maier)
+  - Andreas Scheuring (@scheuran)
+  - Juergen Leopold (@leopoldjuergen)
 requirements:
   - Network access to HMC
   - zhmcclient >=0.14.0
@@ -80,54 +70,57 @@ options:
         required: true
   cpc_name:
     description:
-      - The name of the CPC with the partition containing the virtual function.
+      - The name of the CPC with the partition containing the NIC.
     type: str
     required: true
   partition_name:
     description:
-      - The name of the partition containing the virtual function.
+      - The name of the partition containing the NIC.
     type: str
     required: true
   name:
     description:
-      - The name of the target virtual function that is managed. If the virtual
-        function needs to be created, this value becomes its name.
+      - The name of the target NIC that is managed. If the NIC needs to be
+        created, this value becomes its name.
     type: str
     required: true
   state:
     description:
-      - "The desired state for the target virtual function:"
-      - "C(absent): Ensures that the virtual function does not exist in the
-         specified partition."
-      - "C(present): Ensures that the virtual function exists in the specified
-         partition and has the specified properties."
+      - "The desired state for the target NIC:"
+      - "C(absent): Ensures that the NIC does not exist in the specified
+         partition."
+      - "C(present): Ensures that the NIC exists in the specified partition
+         and has the specified properties."
     type: str
     required: true
     choices: ["absent", "present"]
   properties:
     description:
-      - "Dictionary with input properties for the virtual function, for
-         C(state=present). Key is the property name with underscores instead of
-         hyphens, and value is the property value in YAML syntax. Integer
-         properties may also be provided as decimal strings. Will be ignored
-         for C(state=absent)."
+      - "Dictionary with input properties for the NIC, for C(state=present).
+         Key is the property name with underscores instead of hyphens, and
+         value is the property value in YAML syntax. Integer properties may
+         also be provided as decimal strings. Will be ignored for
+         C(state=absent)."
       - "The possible input properties in this dictionary are the properties
-         defined as writeable in the data model for Virtual Function resources
-         (where the property names contain underscores instead of hyphens),
-         with the following exceptions:"
+         defined as writeable in the data model for NIC resources (where the
+         property names contain underscores instead of hyphens), with the
+         following exceptions:"
       - "* C(name): Cannot be specified because the name has already been
          specified in the C(name) module parameter."
-      - "* C(adapter_uri): Cannot be specified because this information is
-         specified using the artificial property C(adapter_name)."
-      - "* C(adapter_name): The name of the adapter that backs the target
-         virtual function."
+      - "* C(network_adapter_port_uri) and C(virtual_switch_uri): Cannot be
+         specified because this information is specified using the artificial
+         properties C(adapter_name) and C(adapter_port)."
+      - "* C(adapter_name): The name of the adapter that has the port backing
+         the target NIC. Used for all adapter families (ROCE, OSA,
+         Hipersockets)."
+      - "* C(adapter_port): The port index of the adapter port backing the
+         target NIC. Used for all adapter families (ROCE, OSA, Hipersockets)."
       - "Properties omitted in this dictionary will remain unchanged when the
-         virtual function already exists, and will get the default value
-         defined in the data model for virtual functions when the virtual
-         function is being created."
+         NIC already exists, and will get the default value defined in the
+         data model for NICs when the NIC is being created."
     type: dict
     required: false
-    default: No input properties
+    default: null
   log_file:
     description:
       - "File path of a log file to which the logic flow of this module as well
@@ -142,64 +135,89 @@ options:
          If provided, it will be used instead of connecting to a real HMC. This
          is used for testing purposes only."
     required: false
-    default: Real HMC will be used.
+    type: raw
 """
 
 EXAMPLES = """
 ---
 # Note: The following examples assume that some variables named 'my_*' are set.
 
-- name: Ensure virtual function exists in the partition
+- name: Ensure NIC exists in the partition
   zhmc_partition:
     hmc_host: "{{ my_hmc_host }}"
     hmc_auth: "{{ my_hmc_auth }}"
     cpc_name: "{{ my_cpc_name }}"
     partition_name: "{{ my_partition_name }}"
-    name: "{{ my_vfunction_name }}"
+    name: "{{ my_nic_name }}"
     state: present
     properties:
-      adapter_name: "ABC-123"
-      description: "The accelerator adapter"
-      device_number: "033F"
-  register: vfunction1
+      adapter_name: "OSD 0128 A13B-13"
+      adapter_port: 0
+      description: "The port to our data network"
+      device_number: "023F"
+  register: nic1
 
-- name: Ensure virtual function does not exist in the partition
+- name: Ensure NIC does not exist in the partition
   zhmc_partition:
     hmc_host: "{{ my_hmc_host }}"
     hmc_auth: "{{ my_hmc_auth }}"
     cpc_name: "{{ my_cpc_name }}"
     partition_name: "{{ my_partition_name }}"
-    name: "{{ my_vfunction_name }}"
+    name: "{{ my_nic_name }}"
     state: absent
 """
 
 RETURN = """
-virtual_function:
+nic:
   description:
     - "For C(state=absent), an empty dictionary."
     - "For C(state=present), a dictionary with the resource properties of the
-       virtual function (after changes, if any). The dictionary keys are the
-       exact property names as described in the data model for the resource,
-       i.e. they contain hyphens (-), not underscores (_). The dictionary
-       values are the property values using the Python representations
-       described in the documentation of the zhmcclient Python package."
+       NIC (after changes, if any). The dictionary keys are the exact property
+       names as described in the data model for the resource, i.e. they contain
+       hyphens (-), not underscores (_). The dictionary values are the property
+       values using the Python representations described in the documentation
+       of the zhmcclient Python package."
   returned: success
   type: dict
   sample: |
     C({
-      "name": "vfunction-1",
-      "description": "virtual function #1",
-      "adapter-uri': "/api/adapters/...",
+      "name": "nic-1",
+      "description": "NIC #1",
+      "virtual-switch-uri': "/api/vswitches/...",
       ...
     })
 """
 
+import logging  # noqa: E402
+import traceback  # noqa: E402
+from ansible.module_utils.basic import AnsibleModule  # noqa: E402
+
+from ..module_utils.common import log_init, Error, ParameterError, \
+    wait_for_transition_completion, eq_hex, eq_mac, get_hmc_auth, \
+    get_session, to_unicode, process_normal_property, \
+    missing_required_lib  # noqa: E402
+
+try:
+    import requests.packages.urllib3
+    IMP_URLLIB3 = True
+except ImportError:
+    IMP_URLLIB3 = False
+    IMP_URLLIB3_ERR = traceback.format_exc()
+
+try:
+    import zhmcclient
+    IMP_ZHMCCLIENT = True
+except ImportError:
+    IMP_ZHMCCLIENT = False
+    IMP_ZHMCCLIENT_ERR = traceback.format_exc()
+
+
 # Python logger name for this module
-LOGGER_NAME = 'zhmc_virtual_function'
+LOGGER_NAME = 'zhmc_nic'
 
 LOGGER = logging.getLogger(LOGGER_NAME)
 
-# Dictionary of properties of virtual function resources, in this format:
+# Dictionary of properties of NIC resources, in this format:
 #   name: (allowed, create, update, update_while_active, eq_func, type_cast)
 # where:
 #   name: Name of the property according to the data model, with hyphens
@@ -207,42 +225,53 @@ LOGGER = logging.getLogger(LOGGER_NAME)
 #     the 'properties' module parameter).
 #   allowed: Indicates whether it is allowed in the 'properties' module
 #     parameter.
-#   create: Indicates whether it can be specified for the "Create Virtual
-#     Function" operation.
-#   update: Indicates whether it can be specified for the "Update Virtual
-#     Function Properties" operation (at all).
+#   create: Indicates whether it can be specified for the "Create NIC"
+#     operation.
+#   update: Indicates whether it can be specified for the "Update NIC
+#     Properties" operation (at all).
 #   update_while_active: Indicates whether it can be specified for the "Update
-#     Virtual Function Properties" operation while the partition of the
-#     virtual function is active. None means "not applicable" (i.e.
-#     update=False).
+#     NIC Properties" operation while the partition of the NIC is active. None
+#     means "not applicable" (i.e. update=False).
 #   eq_func: Equality test function for two values of the property; None means
 #     to use Python equality.
 #   type_cast: Type cast function for an input value of the property; None
 #     means to use it directly. This can be used for example to convert
 #     integers provided as strings by Ansible back into integers (that is a
 #     current deficiency of Ansible).
-ZHMC_VFUNCTION_PROPERTIES = {
+ZHMC_NIC_PROPERTIES = {
 
     # create+update properties:
     'name': (
         False, True, True, True, None, None),  # provided in 'name' module parm
     'description': (True, True, True, True, None, to_unicode),
     'device_number': (True, True, True, True, eq_hex, None),
-    'adapter_uri': (
-        False, True, True, True, None, None),  # via adapter_name
+    'network_adapter_port_uri': (
+        False, True, True, True, None, None),  # via adapter_name/_port
+    'virtual_switch_uri': (
+        False, True, True, True, None, None),  # via adapter_name/_port
     'adapter_name': (
         True, True, True, True, None,
         None),  # artificial property, type_cast ignored
+    'adapter_port': (
+        True, True, True, True, None,
+        None),  # artificial property, type_cast ignored
+    'ssc_management_nic': (True, True, True, True, None, None),
+    'ssc_ip_address_type': (True, True, True, True, None, None),
+    'ssc_ip_address': (True, True, True, True, None, None),
+    'ssc_mask_prefix': (True, True, True, True, None, None),
+    'vlan_id': (True, True, True, True, None, int),
 
     # read-only properties:
     'element-uri': (False, False, False, None, None, None),
     'element-id': (False, False, False, None, None, None),
     'parent': (False, False, False, None, None, None),
     'class': (False, False, False, None, None, None),
+    'type': (False, False, False, None, None, None),
+    'mac_address': (False, False, False, None, eq_mac, None),
 }
 
 
-def process_properties(partition, vfunction, params):
+def process_properties(partition, nic, params):
     """
     Process the properties specified in the 'properties' module parameter,
     and return two dictionaries (create_props, update_props) that contain
@@ -261,24 +290,22 @@ def process_properties(partition, vfunction, params):
 
     Parameters:
 
-      partition (zhmcclient.Partition): Partition containing the virtual
-        function. Must exist.
+      partition (zhmcclient.Partition): Partition containing the NIC. Must
+        exist.
 
-      vfunction (zhmcclient.VirtualFunction): Virtual function to be updated
-        with the full set of current properties, or `None` if it did not
-        previously exist.
+      nic (zhmcclient.Nic): NIC to be updated with the full set of current
+        properties, or `None` if it did not previously exist.
 
       params (dict): Module input parameters.
 
     Returns:
       tuple of (create_props, update_props, stop), where:
         * create_props: dict of properties for
-          zhmcclient.VirtualFunctionManager.create()
+          zhmcclient.NicManager.create()
         * update_props: dict of properties for
-          zhmcclient.VirtualFunction.update_properties()
+          zhmcclient.Nic.update_properties()
         * stop (bool): Indicates whether some update properties require the
-          partition containg the virtual function to be stopped when doing the
-          update.
+          partition containg the NIC to be stopped when doing the update.
 
     Raises:
       ParameterError: An issue with the module parameters.
@@ -288,13 +315,13 @@ def process_properties(partition, vfunction, params):
     stop = False
 
     # handle 'name' property
-    vfunction_name = to_unicode(params['name'])
-    create_props['name'] = vfunction_name
-    # We looked up the virtual function by name, so we will never have to
-    # update its name
+    nic_name = to_unicode(params['name'])
+    create_props['name'] = nic_name
+    # We looked up the NIC by name, so we will never have to update its name
 
     # Names of the artificial properties
     adapter_name_art_name = 'adapter_name'
+    adapter_port_art_name = 'adapter_port'
 
     # handle the other properties
     input_props = params.get('properties', {})
@@ -302,61 +329,113 @@ def process_properties(partition, vfunction, params):
         input_props = {}
     for prop_name in input_props:
 
-        if prop_name not in ZHMC_VFUNCTION_PROPERTIES:
+        if prop_name not in ZHMC_NIC_PROPERTIES:
             raise ParameterError(
-                "Property {!r} is not defined in the data model for "
-                "virtual functions.".format(prop_name))
+                "Property {0!r} is not defined in the data model for "
+                "NICs.".format(prop_name))
 
         allowed, create, update, update_while_active, eq_func, type_cast = \
-            ZHMC_VFUNCTION_PROPERTIES[prop_name]
+            ZHMC_NIC_PROPERTIES[prop_name]
 
         if not allowed:
             raise ParameterError(
-                "Property {!r} is not allowed in the 'properties' module "
+                "Property {0!r} is not allowed in the 'properties' module "
                 "parameter.".format(prop_name))
 
-        if prop_name == adapter_name_art_name:
+        if prop_name in (adapter_name_art_name, adapter_port_art_name):
             # Artificial properties will be processed together after this loop
             continue
 
         # Process a normal (= non-artificial) property
         _create_props, _update_props, _stop = process_normal_property(
-            prop_name, ZHMC_VFUNCTION_PROPERTIES, input_props, vfunction)
+            prop_name, ZHMC_NIC_PROPERTIES, input_props, nic)
         create_props.update(_create_props)
         update_props.update(_update_props)
         if _stop:
             stop = True
 
     # Process artificial properties
-    if adapter_name_art_name in input_props:
+    if (adapter_name_art_name in input_props) != \
+            (adapter_port_art_name in input_props):
+        raise ParameterError(
+            "Artificial properties {0!r} and {1!r} must either both be "
+            "specified or both be omitted.".
+            format(adapter_name_art_name, adapter_port_art_name))
+    if adapter_name_art_name in input_props and \
+            adapter_port_art_name in input_props:
         adapter_name = to_unicode(input_props[adapter_name_art_name])
+        adapter_port_index = int(input_props[adapter_port_art_name])
         try:
             adapter = partition.manager.cpc.adapters.find(
                 name=adapter_name)
         except zhmcclient.NotFound:
             raise ParameterError(
-                "Artificial property {!r} does not specify the name of an "
-                "existing adapter: {!r}".
+                "Artificial property {0!r} does not specify the name of an "
+                "existing adapter: {1!r}".
                 format(adapter_name_art_name, adapter_name))
+        try:
+            port = adapter.ports.find(index=adapter_port_index)
+        except zhmcclient.NotFound:
+            raise ParameterError(
+                "Artificial property {0!r} does not specify the index of an "
+                "existing port on adapter {1!r}: {2!r}".
+                format(adapter_port_art_name, adapter_name,
+                       adapter_port_index))
 
-        # Here we perform the same logic as in the property loop, just now
-        # simplified by the knowledge about the property flags (create, update,
-        # etc.).
-        hmc_prop_name = 'adapter-uri'
-        input_prop_value = adapter.uri
-        if vfunction:
-            if vfunction.properties.get(hmc_prop_name) != input_prop_value:
+        # The rest of it depends on the network adapter family:
+        adapter_family = adapter.get_property('adapter-family')
+        if adapter_family == 'roce':
+            # Here we perform the same logic as in the property loop, just now
+            # simplified by the knowledge about the property flags (create,
+            # update, etc.).
+            hmc_prop_name = 'network-adapter-port-uri'
+            input_prop_value = port.uri
+            if nic:
+                if nic.properties.get(hmc_prop_name) != input_prop_value:
+                    update_props[hmc_prop_name] = input_prop_value
+            else:
                 update_props[hmc_prop_name] = input_prop_value
+            create_props[hmc_prop_name] = input_prop_value
+        elif adapter_family in ('osa', 'hipersockets'):
+            vswitches = partition.manager.cpc.virtual_switches.findall(
+                **{'backing-adapter-uri': adapter.uri})
+            # Adapters of this family always have a vswitch (one for each
+            # port), so we assert that we can find one or more:
+            if not vswitches:
+                raise AssertionError()
+            found_vswitch = None
+            for vswitch in vswitches:
+                if vswitch.get_property('port') == adapter_port_index:
+                    found_vswitch = vswitch
+                    break
+            # Because we already checked for the existence of the specified
+            # port index, we can now assert that we found the vswitch for that
+            # port:
+            if not found_vswitch:
+                raise AssertionError()
+            # Here we perform the same logic as in the property loop, just now
+            # simplified by the knowledge about the property flags (create,
+            # update, etc.).
+            hmc_prop_name = 'virtual-switch-uri'
+            input_prop_value = found_vswitch.uri
+            if nic:
+                if nic.properties.get(hmc_prop_name) != input_prop_value:
+                    update_props[hmc_prop_name] = input_prop_value
+            else:
+                update_props[hmc_prop_name] = input_prop_value
+            create_props[hmc_prop_name] = input_prop_value
         else:
-            update_props[hmc_prop_name] = input_prop_value
-        create_props[hmc_prop_name] = input_prop_value
+            raise ParameterError(
+                "Artificial property {0!r} specifies the name of a "
+                "non-network adapter of family {1!r}: {2!r}".
+                format(adapter_name_art_name, adapter_family, adapter_name))
 
     return create_props, update_props, stop
 
 
 def ensure_present(params, check_mode):
     """
-    Ensure that the virtual function exists and has the specified properties.
+    Ensure that the NIC exists and has the specified properties.
 
     Raises:
       ParameterError: An issue with the module parameters.
@@ -368,7 +447,7 @@ def ensure_present(params, check_mode):
     userid, password = get_hmc_auth(params['hmc_auth'])
     cpc_name = params['cpc_name']
     partition_name = params['partition_name']
-    vfunction_name = params['name']
+    nic_name = params['name']
     faked_session = params.get('faked_session', None)
 
     changed = False
@@ -384,63 +463,64 @@ def ensure_present(params, check_mode):
             partition = cpc.partitions.find(name=partition_name)
         except zhmcclient.NotFound:
             if check_mode:
-                # Once the partition is created, the virtual function  will
-                # also need to be created. Therefore, we set changed.
+                # Once the partition is created, the NIC will also need to be
+                # created. Therefore, we set changed.
                 changed = True
                 return changed, result
             raise
 
         try:
-            vfunction = partition.virtual_functions.find(name=vfunction_name)
-            vfunction.pull_full_properties()
+            nic = partition.nics.find(name=nic_name)
+            nic.pull_full_properties()
         except zhmcclient.NotFound:
-            vfunction = None
+            nic = None
 
-        if not vfunction:
+        if not nic:
             # It does not exist. Create it and update it if there are
             # update-only properties.
             if not check_mode:
                 create_props, update_props, stop = process_properties(
-                    partition, vfunction, params)
-                vfunction = partition.virtual_functions.create(create_props)
+                    partition, nic, params)
+                nic = partition.nics.create(create_props)
                 update2_props = {}
                 for name in update_props:
                     if name not in create_props:
                         update2_props[name] = update_props[name]
                 if update2_props:
-                    vfunction.update_properties(update2_props)
+                    nic.update_properties(update2_props)
                 # We refresh the properties after the update, in case an
                 # input property value gets changed (for example, the
                 # partition does that with memory properties).
-                vfunction.pull_full_properties()
+                nic.pull_full_properties()
             else:
                 # TODO: Show props in module result also in check mode.
                 pass
             changed = True
         else:
-            # It exists. Stop the partition if needed due to the virtual
-            # function property update requirements, or wait for an updateable
-            # partition status, and update the virtual function properties.
+            # It exists. Stop the partition if needed due to the NIC property
+            # update requirements, or wait for an updateable partition status,
+            # and update the NIC properties.
             create_props, update_props, stop = process_properties(
-                partition, vfunction, params)
+                partition, nic, params)
             if update_props:
                 if not check_mode:
-                    # Virtual function properties can all be updated while the
-                    # partition is active, therefore:
-                    assert not stop
+                    # NIC properties can all be updated while the partition is
+                    # active, therefore:
+                    if stop:
+                        raise AssertionError()
                     wait_for_transition_completion(partition)
-                    vfunction.update_properties(update_props)
+                    nic.update_properties(update_props)
                     # We refresh the properties after the update, in case an
                     # input property value gets changed (for example, the
                     # partition does that with memory properties).
-                    vfunction.pull_full_properties()
+                    nic.pull_full_properties()
                 else:
                     # TODO: Show updated props in mod.result also in chk.mode
                     pass
                 changed = True
 
-        if vfunction:
-            result = vfunction.properties
+        if nic:
+            result = nic.properties
 
         return changed, result
 
@@ -450,7 +530,7 @@ def ensure_present(params, check_mode):
 
 def ensure_absent(params, check_mode):
     """
-    Ensure that the virtual function does not exist.
+    Ensure that the NIC does not exist.
 
     Raises:
       ParameterError: An issue with the module parameters.
@@ -462,7 +542,7 @@ def ensure_absent(params, check_mode):
     userid, password = get_hmc_auth(params['hmc_auth'])
     cpc_name = params['cpc_name']
     partition_name = params['partition_name']
-    vfunction_name = params['name']
+    nic_name = params['name']
     faked_session = params.get('faked_session', None)
 
     changed = False
@@ -476,12 +556,12 @@ def ensure_absent(params, check_mode):
         # The default exception handling is sufficient for the above.
 
         try:
-            vfunction = partition.virtual_functions.find(name=vfunction_name)
+            nic = partition.nics.find(name=nic_name)
         except zhmcclient.NotFound:
             return changed, result
 
         if not check_mode:
-            vfunction.delete()
+            nic.delete()
         changed = True
 
         return changed, result
@@ -524,19 +604,28 @@ def main():
                    choices=['absent', 'present']),
         properties=dict(required=False, type='dict', default={}),
         log_file=dict(required=False, type='str', default=None),
-        faked_session=dict(required=False, type='object'),
+        faked_session=dict(required=False, type='raw'),
     )
 
     module = AnsibleModule(
         argument_spec=argument_spec,
         supports_check_mode=True)
 
+    if not IMP_URLLIB3:
+        module.fail_json(msg=missing_required_lib("requests"),
+                         exception=IMP_URLLIB3_ERR)
+
+    requests.packages.urllib3.disable_warnings()
+
+    if not IMP_ZHMCCLIENT:
+        module.fail_json(msg=missing_required_lib("zhmcclient"),
+                         exception=IMP_ZHMCCLIENT_ERR)
     log_file = module.params['log_file']
     log_init(LOGGER_NAME, log_file)
 
     _params = dict(module.params)
     del _params['hmc_auth']
-    LOGGER.debug("Module entry: params: {!r}".format(_params))
+    LOGGER.debug("Module entry: params: %r", _params)
 
     try:
 
@@ -546,20 +635,17 @@ def main():
         # These exceptions are considered errors in the environment or in user
         # input. They have a proper message that stands on its own, so we
         # simply pass that message on and will not need a traceback.
-        msg = "{}: {}".format(exc.__class__.__name__, exc)
+        msg = "{0}: {1}".format(exc.__class__.__name__, exc)
         LOGGER.debug(
-            "Module exit (failure): msg: {!r}".
-            format(msg))
+            "Module exit (failure): msg: %s", msg)
         module.fail_json(msg=msg)
     # Other exceptions are considered module errors and are handled by Ansible
     # by showing the traceback.
 
     LOGGER.debug(
-        "Module exit (success): changed: {!r}, cpc: {!r}".
-        format(changed, result))
-    module.exit_json(changed=changed, virtual_function=result)
+        "Module exit (success): changed: %r, cpc: %r", changed, result)
+    module.exit_json(changed=changed, nic=result)
 
 
 if __name__ == '__main__':
-    requests.packages.urllib3.disable_warnings()
     main()
