+++ plugins/modules/zhmc_virtual_function.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/python
 # Copyright 2017 IBM Corp. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,16 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import, print_function
-
-import logging
-from ansible.module_utils.basic import AnsibleModule
-import requests.packages.urllib3
-import zhmcclient
-
-from zhmc_ansible_modules.utils import log_init, Error, ParameterError, \
-    wait_for_transition_completion, eq_hex, get_hmc_auth, get_session, \
-    to_unicode, process_normal_property
+from __future__ import (absolute_import, division, print_function)
+__metaclass__ = type
 
 # For information on the format of the ANSIBLE_METADATA, DOCUMENTATION,
 # EXAMPLES, and RETURN strings, see
@@ -38,20 +30,19 @@ ANSIBLE_METADATA = {
 
 DOCUMENTATION = """
 ---
-module: zhmc_hba
-version_added: "0.0"
-short_description: Manages HBAs in existing partitions (without
-    "dpm-storage-management" feature)
+module: zhmc_virtual_function
+short_description: Manages virtual functions in existing partitions
 description:
-  - Creates, updates, and deletes HBAs in existing partitions of a CPC.
+  - Creates, updates, and deletes virtual functions in existing partitions of a
+    CPC.
   - The targeted CPC must be in the Dynamic Partition Manager (DPM) operational
     mode.
 notes:
   - See also Ansible module zhmc_partition.
 author:
-  - Andreas Maier (@andy-maier, maiera@de.ibm.com)
-  - Andreas Scheuring (@scheuran, scheuran@de.ibm.com)
-  - Juergen Leopold (@leopoldjuergen, leopoldj@de.ibm.com)
+  - Andreas Maier (@andy-maier)
+  - Andreas Scheuring (@scheuran)
+  - Juergen Leopold (@leopoldjuergen)
 requirements:
   - Network access to HMC
   - zhmcclient >=0.14.0
@@ -80,56 +71,54 @@ options:
         required: true
   cpc_name:
     description:
-      - The name of the CPC with the partition containing the HBA.
+      - The name of the CPC with the partition containing the virtual function.
     type: str
     required: true
   partition_name:
     description:
-      - The name of the partition containing the HBA.
+      - The name of the partition containing the virtual function.
     type: str
     required: true
   name:
     description:
-      - The name of the target HBA that is managed. If the HBA needs to be
-        created, this value becomes its name.
+      - The name of the target virtual function that is managed. If the virtual
+        function needs to be created, this value becomes its name.
     type: str
     required: true
   state:
     description:
-      - "The desired state for the target HBA:"
-      - "C(absent): Ensures that the HBA does not exist in the specified
-         partition."
-      - "C(present): Ensures that the HBA exists in the specified partition
-         and has the specified properties."
+      - "The desired state for the target virtual function:"
+      - "C(absent): Ensures that the virtual function does not exist in the
+         specified partition."
+      - "C(present): Ensures that the virtual function exists in the specified
+         partition and has the specified properties."
     type: str
     required: true
     choices: ["absent", "present"]
   properties:
     description:
-      - "Dictionary with input properties for the HBA, for C(state=present).
-         Key is the property name with underscores instead of hyphens, and
-         value is the property value in YAML syntax. Integer properties may
-         also be provided as decimal strings. Will be ignored for
-         C(state=absent)."
+      - "Dictionary with input properties for the virtual function, for
+         C(state=present). Key is the property name with underscores instead of
+         hyphens, and value is the property value in YAML syntax. Integer
+         properties may also be provided as decimal strings. Will be ignored
+         for C(state=absent)."
       - "The possible input properties in this dictionary are the properties
-         defined as writeable in the data model for HBA resources (where the
-         property names contain underscores instead of hyphens), with the
-         following exceptions:"
+         defined as writeable in the data model for Virtual Function resources
+         (where the property names contain underscores instead of hyphens),
+         with the following exceptions:"
       - "* C(name): Cannot be specified because the name has already been
          specified in the C(name) module parameter."
-      - "* C(adapter_port_uri): Cannot be specified because this information is
-         specified using the artificial properties C(adapter_name) and
-         C(adapter_port)."
-      - "* C(adapter_name): The name of the adapter that has the port backing
-         the target HBA. Cannot be changed after the HBA exists."
-      - "* C(adapter_port): The port index of the adapter port backing the
-         target HBA. Cannot be changed after the HBA exists."
+      - "* C(adapter_uri): Cannot be specified because this information is
+         specified using the artificial property C(adapter_name)."
+      - "* C(adapter_name): The name of the adapter that backs the target
+         virtual function."
       - "Properties omitted in this dictionary will remain unchanged when the
-         HBA already exists, and will get the default value defined in the
-         data model for HBAs when the HBA is being created."
+         virtual function already exists, and will get the default value
+         defined in the data model for virtual functions when the virtual
+         function is being created."
     type: dict
     required: false
-    default: No input properties
+    default: null
   log_file:
     description:
       - "File path of a log file to which the logic flow of this module as well
@@ -144,65 +133,86 @@ options:
          If provided, it will be used instead of connecting to a real HMC. This
          is used for testing purposes only."
     required: false
-    default: Real HMC will be used.
+    type: raw
 """
 
 EXAMPLES = """
 ---
 # Note: The following examples assume that some variables named 'my_*' are set.
 
-- name: Ensure HBA exists in the partition
+- name: Ensure virtual function exists in the partition
   zhmc_partition:
     hmc_host: "{{ my_hmc_host }}"
     hmc_auth: "{{ my_hmc_auth }}"
     cpc_name: "{{ my_cpc_name }}"
     partition_name: "{{ my_partition_name }}"
-    name: "{{ my_hba_name }}"
+    name: "{{ my_vfunction_name }}"
     state: present
     properties:
-      adapter_name: FCP-1
-      adapter_port: 0
-      description: "The port to our V7K #1"
-      device_number: "123F"
-  register: hba1
+      adapter_name: "ABC-123"
+      description: "The accelerator adapter"
+      device_number: "033F"
+  register: vfunction1
 
-- name: Ensure HBA does not exist in the partition
+- name: Ensure virtual function does not exist in the partition
   zhmc_partition:
     hmc_host: "{{ my_hmc_host }}"
     hmc_auth: "{{ my_hmc_auth }}"
     cpc_name: "{{ my_cpc_name }}"
     partition_name: "{{ my_partition_name }}"
-    name: "{{ my_hba_name }}"
+    name: "{{ my_vfunction_name }}"
     state: absent
 """
 
 RETURN = """
-hba:
+virtual_function:
   description:
     - "For C(state=absent), an empty dictionary."
     - "For C(state=present), a dictionary with the resource properties of the
-       HBA (after changes, if any). The dictionary keys are the exact property
-       names as described in the data model for the resource, i.e. they contain
-       hyphens (-), not underscores (_). The dictionary values are the property
-       values using the Python representations described in the documentation
-       of the zhmcclient Python package."
+       virtual function (after changes, if any). The dictionary keys are the
+       exact property names as described in the data model for the resource,
+       i.e. they contain hyphens (-), not underscores (_). The dictionary
+       values are the property values using the Python representations
+       described in the documentation of the zhmcclient Python package."
   returned: success
   type: dict
   sample: |
     C({
-      "name": "hba-1",
-      "description": "HBA #1",
-      "adapter-port-uri": "/api/adapters/.../ports/...",
+      "name": "vfunction-1",
+      "description": "virtual function #1",
+      "adapter-uri': "/api/adapters/...",
       ...
     })
 """
 
+import logging  # noqa: E402
+import traceback  # noqa: E402
+from ansible.module_utils.basic import AnsibleModule  # noqa: E402
+
+from ..module_utils.common import log_init, Error, ParameterError, \
+    wait_for_transition_completion, eq_hex, get_hmc_auth, get_session, \
+    to_unicode, process_normal_property, missing_required_lib  # noqa: E402
+
+try:
+    import requests.packages.urllib3
+    IMP_URLLIB3 = True
+except ImportError:
+    IMP_URLLIB3 = False
+    IMP_URLLIB3_ERR = traceback.format_exc()
+
+try:
+    import zhmcclient
+    IMP_ZHMCCLIENT = True
+except ImportError:
+    IMP_ZHMCCLIENT = False
+    IMP_ZHMCCLIENT_ERR = traceback.format_exc()
+
 # Python logger name for this module
-LOGGER_NAME = 'zhmc_hba'
+LOGGER_NAME = 'zhmc_virtual_function'
 
 LOGGER = logging.getLogger(LOGGER_NAME)
 
-# Dictionary of properties of HBA resources, in this format:
+# Dictionary of properties of virtual function resources, in this format:
 #   name: (allowed, create, update, update_while_active, eq_func, type_cast)
 # where:
 #   name: Name of the property according to the data model, with hyphens
@@ -210,47 +220,42 @@ LOGGER = logging.getLogger(LOGGER_NAME)
 #     the 'properties' module parameter).
 #   allowed: Indicates whether it is allowed in the 'properties' module
 #     parameter.
-#   create: Indicates whether it can be specified for the "Create HBA"
-#     operation.
-#   update: Indicates whether it can be specified for the "Update HBA
-#     Properties" operation (at all).
+#   create: Indicates whether it can be specified for the "Create Virtual
+#     Function" operation.
+#   update: Indicates whether it can be specified for the "Update Virtual
+#     Function Properties" operation (at all).
 #   update_while_active: Indicates whether it can be specified for the "Update
-#     HBA Properties" operation while the partition of the HBA is active. None
-#     means "not applicable" (i.e. update=False).
+#     Virtual Function Properties" operation while the partition of the
+#     virtual function is active. None means "not applicable" (i.e.
+#     update=False).
 #   eq_func: Equality test function for two values of the property; None means
 #     to use Python equality.
 #   type_cast: Type cast function for an input value of the property; None
 #     means to use it directly. This can be used for example to convert
 #     integers provided as strings by Ansible back into integers (that is a
 #     current deficiency of Ansible).
-ZHMC_HBA_PROPERTIES = {
-
-    # create-only properties:
-    'adapter_port_uri': (
-        False, True, False, None, None, None),  # via adapter_name/_port
-    'adapter_name': (
-        True, True, False, None, None,
-        None),  # artificial property, type_cast ignored
-    'adapter_port': (
-        True, True, False, None, None,
-        None),  # artificial property, type_cast ignored
+ZHMC_VFUNCTION_PROPERTIES = {
 
     # create+update properties:
     'name': (
         False, True, True, True, None, None),  # provided in 'name' module parm
     'description': (True, True, True, True, None, to_unicode),
     'device_number': (True, True, True, True, eq_hex, None),
+    'adapter_uri': (
+        False, True, True, True, None, None),  # via adapter_name
+    'adapter_name': (
+        True, True, True, True, None,
+        None),  # artificial property, type_cast ignored
 
     # read-only properties:
     'element-uri': (False, False, False, None, None, None),
     'element-id': (False, False, False, None, None, None),
     'parent': (False, False, False, None, None, None),
     'class': (False, False, False, None, None, None),
-    'wwpn': (False, False, False, None, None, None),
 }
 
 
-def process_properties(partition, hba, params):
+def process_properties(partition, vfunction, params):
     """
     Process the properties specified in the 'properties' module parameter,
     and return two dictionaries (create_props, update_props) that contain
@@ -269,22 +274,24 @@ def process_properties(partition, hba, params):
 
     Parameters:
 
-      partition (zhmcclient.Partition): Partition containing the HBA. Must
-        exist.
+      partition (zhmcclient.Partition): Partition containing the virtual
+        function. Must exist.
 
-      hba (zhmcclient.Hba): HBA to be updated with the full set of current
-        properties, or `None` if it did not previously exist.
+      vfunction (zhmcclient.VirtualFunction): Virtual function to be updated
+        with the full set of current properties, or `None` if it did not
+        previously exist.
 
       params (dict): Module input parameters.
 
     Returns:
       tuple of (create_props, update_props, stop), where:
         * create_props: dict of properties for
-          zhmcclient.HbaManager.create()
+          zhmcclient.VirtualFunctionManager.create()
         * update_props: dict of properties for
-          zhmcclient.Hba.update_properties()
+          zhmcclient.VirtualFunction.update_properties()
         * stop (bool): Indicates whether some update properties require the
-          partition containg the HBA to be stopped when doing the update.
+          partition containg the virtual function to be stopped when doing the
+          update.
 
     Raises:
       ParameterError: An issue with the module parameters.
@@ -294,13 +301,13 @@ def process_properties(partition, hba, params):
     stop = False
 
     # handle 'name' property
-    hba_name = to_unicode(params['name'])
-    create_props['name'] = hba_name
-    # We looked up the HBA by name, so we will never have to update its name
+    vfunction_name = to_unicode(params['name'])
+    create_props['name'] = vfunction_name
+    # We looked up the virtual function by name, so we will never have to
+    # update its name
 
     # Names of the artificial properties
     adapter_name_art_name = 'adapter_name'
-    adapter_port_art_name = 'adapter_port'
 
     # handle the other properties
     input_props = params.get('properties', {})
@@ -308,74 +315,61 @@ def process_properties(partition, hba, params):
         input_props = {}
     for prop_name in input_props:
 
-        if prop_name not in ZHMC_HBA_PROPERTIES:
+        if prop_name not in ZHMC_VFUNCTION_PROPERTIES:
             raise ParameterError(
-                "Property {!r} is not defined in the data model for "
-                "HBAs.".format(prop_name))
+                "Property {0!r} is not defined in the data model for "
+                "virtual functions.".format(prop_name))
 
         allowed, create, update, update_while_active, eq_func, type_cast = \
-            ZHMC_HBA_PROPERTIES[prop_name]
+            ZHMC_VFUNCTION_PROPERTIES[prop_name]
 
         if not allowed:
             raise ParameterError(
-                "Property {!r} is not allowed in the 'properties' module "
+                "Property {0!r} is not allowed in the 'properties' module "
                 "parameter.".format(prop_name))
 
-        if prop_name in (adapter_name_art_name, adapter_port_art_name):
+        if prop_name == adapter_name_art_name:
             # Artificial properties will be processed together after this loop
             continue
 
         # Process a normal (= non-artificial) property
         _create_props, _update_props, _stop = process_normal_property(
-            prop_name, ZHMC_HBA_PROPERTIES, input_props, hba)
+            prop_name, ZHMC_VFUNCTION_PROPERTIES, input_props, vfunction)
         create_props.update(_create_props)
         update_props.update(_update_props)
         if _stop:
             stop = True
 
     # Process artificial properties
-    if (adapter_name_art_name in input_props) != \
-            (adapter_port_art_name in input_props):
-        raise ParameterError(
-            "Artificial properties {!r} and {!r} must either both be "
-            "specified or both be omitted.".
-            format(adapter_name_art_name, adapter_port_art_name))
-    if adapter_name_art_name in input_props and \
-            adapter_port_art_name in input_props:
+    if adapter_name_art_name in input_props:
         adapter_name = to_unicode(input_props[adapter_name_art_name])
-        adapter_port_index = int(input_props[adapter_port_art_name])
         try:
             adapter = partition.manager.cpc.adapters.find(
                 name=adapter_name)
         except zhmcclient.NotFound:
             raise ParameterError(
-                "Artificial property {!r} does not specify the name of an "
-                "existing adapter: {!r}".
+                "Artificial property {0!r} does not specify the name of an "
+                "existing adapter: {1!r}".
                 format(adapter_name_art_name, adapter_name))
-        try:
-            port = adapter.ports.find(index=adapter_port_index)
-        except zhmcclient.NotFound:
-            raise ParameterError(
-                "Artificial property {!r} does not specify the index of an "
-                "existing port on adapter {!r}: {!r}".
-                format(adapter_port_art_name, adapter_name,
-                       adapter_port_index))
-        hmc_prop_name = 'adapter-port-uri'
-        if hba:
-            existing_port_uri = hba.get_property(hmc_prop_name)
-            if port.uri != existing_port_uri:
-                raise ParameterError(
-                    "Artificial properties {!r} and {!r} cannot be used to "
-                    "change the adapter port of an existing HBA".
-                    format(adapter_name_art_name, adapter_port_art_name))
-        create_props[hmc_prop_name] = port.uri
+
+        # Here we perform the same logic as in the property loop, just now
+        # simplified by the knowledge about the property flags (create, update,
+        # etc.).
+        hmc_prop_name = 'adapter-uri'
+        input_prop_value = adapter.uri
+        if vfunction:
+            if vfunction.properties.get(hmc_prop_name) != input_prop_value:
+                update_props[hmc_prop_name] = input_prop_value
+        else:
+            update_props[hmc_prop_name] = input_prop_value
+        create_props[hmc_prop_name] = input_prop_value
 
     return create_props, update_props, stop
 
 
 def ensure_present(params, check_mode):
     """
-    Ensure that the HBA exists and has the specified properties.
+    Ensure that the virtual function exists and has the specified properties.
 
     Raises:
       ParameterError: An issue with the module parameters.
@@ -387,7 +381,7 @@ def ensure_present(params, check_mode):
     userid, password = get_hmc_auth(params['hmc_auth'])
     cpc_name = params['cpc_name']
     partition_name = params['partition_name']
-    hba_name = params['name']
+    vfunction_name = params['name']
     faked_session = params.get('faked_session', None)
 
     changed = False
@@ -403,63 +397,64 @@ def ensure_present(params, check_mode):
             partition = cpc.partitions.find(name=partition_name)
         except zhmcclient.NotFound:
             if check_mode:
-                # Once the partition is created, the HBA will also need to be
-                # created. Therefore, we set changed.
+                # Once the partition is created, the virtual function  will
+                # also need to be created. Therefore, we set changed.
                 changed = True
                 return changed, result
             raise
 
         try:
-            hba = partition.hbas.find(name=hba_name)
-            hba.pull_full_properties()
+            vfunction = partition.virtual_functions.find(name=vfunction_name)
+            vfunction.pull_full_properties()
         except zhmcclient.NotFound:
-            hba = None
+            vfunction = None
 
-        if not hba:
+        if not vfunction:
             # It does not exist. Create it and update it if there are
             # update-only properties.
             if not check_mode:
                 create_props, update_props, stop = process_properties(
-                    partition, hba, params)
-                hba = partition.hbas.create(create_props)
+                    partition, vfunction, params)
+                vfunction = partition.virtual_functions.create(create_props)
                 update2_props = {}
                 for name in update_props:
                     if name not in create_props:
                         update2_props[name] = update_props[name]
                 if update2_props:
-                    hba.update_properties(update2_props)
+                    vfunction.update_properties(update2_props)
                 # We refresh the properties after the update, in case an
                 # input property value gets changed (for example, the
                 # partition does that with memory properties).
-                hba.pull_full_properties()
+                vfunction.pull_full_properties()
             else:
                 # TODO: Show props in module result also in check mode.
                 pass
             changed = True
         else:
-            # It exists. Stop the partition if needed due to the HBA property
-            # update requirements, or wait for an updateable partition status,
-            # and update the HBA properties.
+            # It exists. Stop the partition if needed due to the virtual
+            # function property update requirements, or wait for an updateable
+            # partition status, and update the virtual function properties.
             create_props, update_props, stop = process_properties(
-                partition, hba, params)
+                partition, vfunction, params)
             if update_props:
                 if not check_mode:
-                    # HBA properties can all be updated while the partition is
-                    # active, therefore:
-                    assert not stop
+                    # Virtual function properties can all be updated while the
+                    # partition is active, therefore:
+                    if stop:
+                        raise AssertionError()
                     wait_for_transition_completion(partition)
-                    hba.update_properties(update_props)
+                    vfunction.update_properties(update_props)
                     # We refresh the properties after the update, in case an
                     # input property value gets changed (for example, the
                     # partition does that with memory properties).
-                    hba.pull_full_properties()
+                    vfunction.pull_full_properties()
                 else:
                     # TODO: Show updated props in mod.result also in chk.mode
                     pass
                 changed = True
 
-        if hba:
-            result = hba.properties
+        if vfunction:
+            result = vfunction.properties
 
         return changed, result
 
@@ -469,7 +464,7 @@ def ensure_present(params, check_mode):
 
 def ensure_absent(params, check_mode):
     """
-    Ensure that the HBA does not exist.
+    Ensure that the virtual function does not exist.
 
     Raises:
       ParameterError: An issue with the module parameters.
@@ -481,7 +476,7 @@ def ensure_absent(params, check_mode):
     userid, password = get_hmc_auth(params['hmc_auth'])
     cpc_name = params['cpc_name']
     partition_name = params['partition_name']
-    hba_name = params['name']
+    vfunction_name = params['name']
     faked_session = params.get('faked_session', None)
 
     changed = False
@@ -495,12 +490,12 @@ def ensure_absent(params, check_mode):
         # The default exception handling is sufficient for the above.
 
         try:
-            hba = partition.hbas.find(name=hba_name)
+            vfunction = partition.virtual_functions.find(name=vfunction_name)
         except zhmcclient.NotFound:
             return changed, result
 
         if not check_mode:
-            hba.delete()
+            vfunction.delete()
         changed = True
 
         return changed, result
@@ -543,19 +538,29 @@ def main():
                    choices=['absent', 'present']),
         properties=dict(required=False, type='dict', default={}),
         log_file=dict(required=False, type='str', default=None),
-        faked_session=dict(required=False, type='object'),
+        faked_session=dict(required=False, type='raw'),
     )
 
     module = AnsibleModule(
         argument_spec=argument_spec,
         supports_check_mode=True)
 
+    if not IMP_URLLIB3:
+        module.fail_json(msg=missing_required_lib("requests"),
+                         exception=IMP_URLLIB3_ERR)
+
+    requests.packages.urllib3.disable_warnings()
+
+    if not IMP_ZHMCCLIENT:
+        module.fail_json(msg=missing_required_lib("zhmcclient"),
+                         exception=IMP_ZHMCCLIENT_ERR)
+
     log_file = module.params['log_file']
     log_init(LOGGER_NAME, log_file)
 
     _params = dict(module.params)
     del _params['hmc_auth']
-    LOGGER.debug("Module entry: params: {!r}".format(_params))
+    LOGGER.debug("Module entry: params: %r", _params)
 
     try:
 
@@ -565,20 +570,17 @@ def main():
         # These exceptions are considered errors in the environment or in user
         # input. They have a proper message that stands on its own, so we
         # simply pass that message on and will not need a traceback.
-        msg = "{}: {}".format(exc.__class__.__name__, exc)
+        msg = "{0}: {1}".format(exc.__class__.__name__, exc)
         LOGGER.debug(
-            "Module exit (failure): msg: {!r}".
-            format(msg))
+            "Module exit (failure): msg: %s", msg)
         module.fail_json(msg=msg)
     # Other exceptions are considered module errors and are handled by Ansible
     # by showing the traceback.
 
     LOGGER.debug(
-        "Module exit (success): changed: {!r}, cpc: {!r}".
-        format(changed, result))
-    module.exit_json(changed=changed, hba=result)
+        "Module exit (success): changed: %r, cpc: %s", changed, result)
+    module.exit_json(changed=changed, virtual_function=result)
 
 
 if __name__ == '__main__':
-    requests.packages.urllib3.disable_warnings()
     main()
