+++ tests/function/__init__.py
@@ -1,0 +1,1 @@
+# this file is required to get the pytest working with relative imports
