+++ tests/sanity/ignore-2.9.txt
@@ -1,0 +1,14 @@
+plugins/modules/zhmc_adapter.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_crypto_attachment.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_cpc.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_hba.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_nic.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_partition.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_storage_group.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_storage_group_attachment.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_storage_volume.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_user.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+plugins/modules/zhmc_virtual_function.py validate-modules:missing-gplv3-license # Licensed under Apache 2.0
+os_setup.sh shebang # using /bin/bash
+remove_duplicate_setuptools.py shebang #using /usr/bin/python
+setup.py pylint:invalid-name # the __metaclass__ name cannot be changed
